general cleanup from the current refactor (#831)
* general cleanup

* more small changes
