+++ conf/lambda.json
@@ -55,7 +55,6 @@
       "PREFIX_GOES_HERE.streamalerts": "alerts"
     },
     "concurrency_limit": 10,
-    "enable_metrics": false,
     "log_level": "info"
   },
   "rule_processor_config": {
