+++ tests/unit/conf/threat_intel.json
@@ -1,0 +1,17 @@
+{
+  "dynamodb_table_name": "table_name",
+  "enabled": false,
+  "excluded_iocs": {
+    "ip": {
+      "10.0.0.0/8": {
+        "comment": "RFC1918"
+      }
+    }
+  },
+  "normalized_ioc_types": {
+    "ip": [
+      "sourceAddress",
+      "destinationAddress"
+    ]
+  }
+}
